Count and normalize (and transform) ChIP-seq read coverage

Count ChIP-seq read coverage for each bam file using the coverage function from bedtools, Then normalize the read counts by scaling to the reference ChIP-seq library size, and transform the ChIP-seq read counts.

count_normalize_chip(
  sites.df,
  chip_bam_files,
  chip_idxstats_files,
  chrom_size_file,
  ref.size = 1e+07,
  transform = c("none", "asinh", "log2", "sqrt"),
  bedtools_path = "bedtools"
)

Arguments

sites.df

A data frame containing the candidate sites.

chip_bam_files

ChIP-seq bam files (may include multiple replicates).

chip_idxstats_files

ChIP-seq idxstats files (may include multiple replicates). (Default: use the corresponding '.idxstats.txt' files in the same directory of the bam files).

chrom_size_file

Chromosome size file.

ref.size

Reference ChIP-seq library size (Default: 10 million)

transform

Type of transformation for the ChIP read counts. Options are 'none' (no transformation), 'asinh', 'log2', and 'sqrt'.

bedtools_path

Path to bedtools executable.

Value

A data frame of candidate sites and normalized (and transformed) ChIP-seq read counts around each candidate site.