Run FIMO to scan for motif matches

Run FIMO to scan for motif matches along the genome. This is a wrapper function to run FIMO command line.

fimo_motif_matches(
  motif_file,
  sequence_file,
  outname = "fimo.txt",
  outdir = dirname(outname),
  thresh_pValue = 1e-04,
  bfile,
  skip_matched_sequence = TRUE,
  max_strand = FALSE,
  options = "",
  verbosity = c(1, 2, 3, 4, 5),
  fimo_path = "fimo"
)

Arguments

motif_file

Motif file.

sequence_file

FASTA sequence file.

outname

Output file name.

outdir

Output directory.

thresh_pValue

FIMO option 'thresh' for p-value threshold.

bfile

FIMO option 'bfile' for background file or model.

skip_matched_sequence

FIMO option 'skip_matched_sequence'. Turns off output of the sequence of motif matches. This speeds up processing considerably.

max_strand

FIMO option 'max_strand'. If matches on both strands at a given position satisfy the output threshold, only report the match for the strand with the higher score. If the scores are tied, the matching strand is chosen at random.

options

Other options for FIMO.

verbosity

A number that regulates the verbosity level of the output information messages. If set to 1 (quiet) then it will only output error messages whereas the other extreme 5 (dump) outputs lots of mostly useless information.

fimo_path

Path to FIMO command line executable.