Fit TOP model

Fit TOP model (using M5 bins) for selected partitions in parallel.

fit_TOP_M5_model(
  all_training_data,
  all_training_data_files,
  model_file,
  logistic_model = FALSE,
  transform = c("asinh", "log2", "sqrt", "none"),
  partitions = 1:10,
  n_iter = 2000,
  n_burnin = floor(n_iter/2),
  n_chains = 3,
  n_thin = max(1, floor((n_iter - n_burnin)/1000)),
  n_cores = length(partitions),
  save = TRUE,
  outdir = "./TOP_samples",
  return_type = c("samples", "jagsfit", "samplefiles"),
  quiet = FALSE
)

Arguments

all_training_data

A list of the assembled training data of all partitions.

all_training_data_files

A vector of the assembled training data files of all partitions. If all_training_data is missing, it will load the training data from all_training_data_files.

model_file

TOP model file written in JAGS.

logistic_model

Logical; if TRUE, use the logistic version of TOP model.

transform

Type of transformation for ChIP counts. Possible values are "asinh", "log2", "sqrt", and "none" (no transformation). Only needed when logistic_model is FALSE.

partitions

A vector of selected partition(s) to run. (default: all 10 partitions (1:10))

n_iter

Number of total iterations per chain (including burn in).

n_burnin

Length of burn in samples, i.e. number of iterations to discard at the beginning. Default is n_iter/2, that is, discarding the first half of the simulations.

n_chains

Number of Markov chains (default: 3).

n_thin

Thinning rate, must be a positive integer. Default is max(1, floor(n_chains * (n_iter-n_burnin) / 1000)) which will only thin if there are at least 2000 simulations.

n_cores

Number of cores to use in parallel (default: equal to the number of partitions).

save

Logical, if TRUE, save posterior samples as .rds files in outdir.

outdir

Directory to save TOP model posterior samples.

return_type

Type of result to return. Options: 'samples' (posterior samples),

quiet

Logical, whether to suppress stdout in jags.model().

Value

A list of posterior samples for each of the partitions. A list of posterior samples or jagsfit object for each partition.