Perform MILLIPEDE binning on count matrix

Perform binning using different MILLIPEDE binning schemes (M5, M24, M12, M3, M2, M1) on the input count matrix.

millipede_binning(
  counts,
  bin_method = c("M5", "M24", "M12", "M3", "M2", "M1"),
  combine_strands = c("vertical", "motif")
)

Arguments

counts

DNase-seq or ATAC-seq read counts matrix, rows are candidate sites, columns are DNase or ATAC counts with 100bp flanks around motifs on the forward and reverse strands.

bin_method

MILLIPEDE binning scheme. Options: “M5” (default), “M24”, “M12”, “M3”, “M2”, and “M1”.

combine_strands

Method to combine counts on both strands from M24 bins to M12 bins: vertical' (default) or 'motif'.

Value

A list containing binning results (data frames) using different binning schemes.

Examples

# Perform MILLIPEDE binning with different binning schemes.

# M5 binning
 M5_bins <- millipede_binning(counts, bin_method = 'M5')
#> Error in as.matrix(counts): object 'counts' not found

# M12 binning
 M12_bins <- millipede_binning(counts, bin_method = 'M12')
#> Error in as.matrix(counts): object 'counts' not found

# M3 binning
 M3_bins <- millipede_binning(counts, bin_method = 'M3')
#> Error in as.matrix(counts): object 'counts' not found