Predict quantitative TF occupancy or TF binding probability

Predict quantitative TF occupancy or TF binding probability using TOP model trained from ChIP-seq read counts or binary labels.

predict_TOP(
  data,
  TOP_coef,
  tf_name,
  cell_type,
  level = c("best", "bottom", "middle", "top"),
  logistic_model = FALSE,
  transform = c("asinh", "log2", "log", "none")
)

Arguments

data

A data frame containing motif PWM score and DNase (or ATAC) bins.

TOP_coef

A list containing the posterior mean of TOP regression coefficients.

tf_name

specifies the TF name to make predictions. It will find the model parameters trained for this TF. This is not needed (not used) when level = 'top'.

cell_type

specifies the cell type to make predictions. It will find the model parameters trained for this cell type. This is not needed (not used) when level = 'middle' or level = 'top'.

level

Specific the TOP model hierarchy level to use. Options: ‘best’, ‘bottom’, ‘middle’, or ‘top’. When level = 'best', use the best (lowest available) level of the hierarchy for the TF x cell type combination. When level = 'bottom', use the bottom level (TF- and cell-type-specific model), if the TF motif and cell type is available in the training data. When level = 'middle', use the middle level (TF-specific model) of that TF. When level = 'top', use the top level TF-generic model.

logistic_model

Logical. Whether to use the logistic version of TOP model. If logistic_model = TRUE, use the logistic version of TOP model to predict TF binding probability. If logistic_model = FALSE, use the quantitative occupancy model (default).

transform

Type of transformation performed for ChIP-seq read counts when preparing the input training data. Options are: ‘asinh’(asinh transformation), ‘log2’ (log2 transformation), ‘sqrt’ (square root transformation), and ‘none’(no transformation). This only applies when logistic_model = FALSE.

Value

Returns a list with the following elements,

model

TOP model name.

level

selected hierarchy level.

coef

posterior mean of regression coefficients.

predictions

a data frame with the data and predicted values.

Examples

# 'data' is a data frame of input data,
# with columns of PWM scores and five DNase (or ATAC) bins.
# 'TOP_coef' is pretrained posterior mean of TOP regression coefficients.

# Predict CTCF occupancy in K562 using the quantitative occupancy model:

# Predict using the 'bottom' level model
result <- predict_TOP(data, TOP_coef,
                      tf_name = 'CTCF', cell_type = 'K562',
                      level = 'bottom',
                      logistic_model = FALSE,
                      transform = 'asinh') # We used 'asinh' transformation on the ChIP data when training the model.
#> Error in select_model_coef_level(TOP_coef, tf_name, cell_type, level): object 'TOP_coef' not found

# Predict using the 'best' model
# Since CTCF in K562 cell type is included in training,
# the 'best' model is the 'bottom' level model.
result <- predict_TOP(data, TOP_coef,
                      tf_name = 'CTCF', cell_type = 'K562', level = 'best',
                      logistic_model = FALSE, transform = 'asinh')
#> Error in select_model_coef_level(TOP_coef, tf_name, cell_type, level): object 'TOP_coef' not found

# We can use the 'middle' model to predict CTCF in K562
# or other cell types or conditions
result <- predict_TOP(data, TOP_coef,
                      tf_name = 'CTCF', level = 'middle',
                      logistic_model = FALSE, transform = 'asinh')
#> Error in select_model_coef_level(TOP_coef, tf_name, cell_type, level): object 'TOP_coef' not found

# Predict CTCF binding probability using the logistic version of the model:
# No need to set the argument for 'transform' for the logistic model.

# Predict using the 'bottom' level model
result <- predict_TOP(data, TOP_coef,
                      tf_name = 'CTCF', cell_type = 'K562',
                      level = 'best',
                      logistic_model = TRUE)
#> Error in select_model_coef_level(TOP_coef, tf_name, cell_type, level): object 'TOP_coef' not found

# Predict using the 'middle' level model
result <- predict_TOP(data, TOP_coef,
                      tf_name = 'CTCF', level = 'middle',
                      logistic_model = TRUE)
#> Error in select_model_coef_level(TOP_coef, tf_name, cell_type, level): object 'TOP_coef' not found