Predict TF occupancy using posterior samples of regression coefficients

Predict TF occupancy using posterior samples of regression coefficients trained from the TOP model

Usage

predict_TOP_samples(
  data,
  TOP_samples,
  use.posterior.mean = FALSE,
  sample.predictions = TRUE,
  transform = c("asinh", "log2", "log", "none")
)

Arguments

data

A data frame. Columns are motif score and DNase (or ATAC) bins. Rows are candidate sites.

TOP_samples

TOP posterior samples.

use.posterior.mean

If TRUE, uses the posterior mean of regression coefficients to make predictions.

sample.predictions

If TRUE, sample from posterior predictions and then take the mean of posterior prediction samples.

transform

Method used to transform ChIP-seq counts when training the TOP model. Options: asinh, log2, log, none.

Value

The function returns a vector of predicted TF occupancy.